src/Functions/requirements.txt
requests
boto3
Pillow
ydb